+++ tst/ruCubeIgnoredPiecesGapFiller_test.cpp
@@ -191,7 +191,7 @@ TEST(ruCubeIgnoredPiecesGapFiller, permutationIgnoredGapsFillNextTest) {
         auto& [cp, ep] = permutations[i];
         filler.permutationIgnoredGapsFillInit(cp, ep);
 
-        uint j = 0;
+        uint16_t j = 0;
         while (filler.permutationIgnoredGapsFillNext(cp, ep)) {
             const auto& [ecp, eep] = expectedPermutations[i][j];
 //            std::cout << "---------------i = " << (int) i << ": " << (int)j << std::endl;
